+++ eras/alonzo/test-suite/src/Test/Cardano/Ledger/Alonzo/Serialisation/Generators.hs
@@ -14,16 +14,8 @@
 module Test.Cardano.Ledger.Alonzo.Serialisation.Generators where
 
 import Cardano.Binary (ToCBOR (..))
-import Cardano.Ledger.Alonzo.Data
-  ( AlonzoAuxiliaryData (..),
-    BinaryData,
-    Data (..),
-    Datum (..),
-    dataToBinaryData,
-  )
-import Cardano.Ledger.Alonzo.Language (Language (..), nonNativeLanguages)
 import Cardano.Ledger.Alonzo (AlonzoEra)
-import Cardano.Ledger.Alonzo.Data (AlonzoAuxiliaryData (..), BinaryData, Data (..), dataToBinaryData, AuxiliaryDataHash)
+import Cardano.Ledger.Alonzo.Data (AlonzoAuxiliaryData (..), AuxiliaryDataHash, BinaryData, Data (..), Datum (..), dataToBinaryData)
 import Cardano.Ledger.Alonzo.Language
 import Cardano.Ledger.Alonzo.PParams
   ( AlonzoPParams,
@@ -48,27 +40,29 @@ import Cardano.Ledger.Alonzo.Scripts
   )
 import Cardano.Ledger.Alonzo.Tx
   ( AlonzoTx (AlonzoTx),
-    AlonzoTxBody (AlonzoTxBody),
+    AlonzoTxBody (..),
     CostModel,
     IsValid (..),
     ScriptIntegrity (..),
     ScriptPurpose (..),
     hashData,
   )
-import Cardano.Ledger.Alonzo.TxBody (AlonzoTxOut (..))
+import Cardano.Ledger.Alonzo.TxBody (AlonzoTxOut (..), ScriptIntegrityHash)
 import Cardano.Ledger.Alonzo.TxWits
-import Cardano.Ledger.Alonzo.TxWitness
-import Cardano.Ledger.BaseTypes (StrictMaybe (..), Network)
+import Cardano.Ledger.BaseTypes (Network, StrictMaybe (..))
 import Cardano.Ledger.Coin (Coin)
 import Cardano.Ledger.Core
-import Control.State.Transition (PredicateFailure)
 import qualified Cardano.Ledger.Crypto as C
+import Cardano.Ledger.Keys (KeyHash)
+import Cardano.Ledger.Mary.Value (MultiAsset)
+import Cardano.Ledger.Shelley.PParams (Update)
 import Cardano.Ledger.Shelley.TxBody (DCert, Wdrl)
 import Cardano.Ledger.ShelleyMA.Timelocks (ValidityInterval (..))
 import Cardano.Ledger.TxIn (TxIn)
-import Cardano.Ledger.Val (DecodeNonNegative, Val, EncodeMint (..))
+import Cardano.Ledger.Val (DecodeNonNegative, EncodeMint (..), Val)
 import Cardano.Slotting.Slot (SlotNo)
-import Codec.CBOR.Term (Term (..), encodeTerm)
+import Codec.CBOR.Term (Term (..))
+import Control.State.Transition (PredicateFailure)
 import Data.Int (Int64)
 import Data.List.NonEmpty (NonEmpty (..))
 import Data.Map (Map)
@@ -77,9 +71,10 @@ import Data.Maybe (catMaybes)
 import Data.Set (Set)
 import qualified Data.Set as Set
 import Data.Text (Text, pack)
-import GHC.Records (HasField)
 import Data.Twiddle
+import Data.Typeable (Typeable)
 import Data.Void (Void)
+import GHC.Records (HasField)
 import Numeric.Natural (Natural)
 import qualified Plutus.V1.Ledger.Api as PV1
 import qualified Plutus.V2.Ledger.Api as PV2
@@ -88,12 +83,6 @@ import Test.Cardano.Ledger.Shelley.ConcreteCryptoTypes (Mock)
 import Test.Cardano.Ledger.Shelley.Serialisation.EraIndepGenerators ()
 import Test.Cardano.Ledger.ShelleyMA.Serialisation.Generators (genMintValues)
 import Test.QuickCheck
-import Cardano.Ledger.Shelley.PParams (Update)
-import Cardano.Ledger.Mary.Value (MultiAsset)
-import Cardano.Ledger.Keys (KeyHash)
-import Data.Typeable (Typeable)
-import Debug.Trace (traceM)
-import Codec.CBOR.Write (toLazyByteString)
 
 instance Era era => Arbitrary (Data era) where
   arbitrary = Data <$> arbitrary
@@ -106,14 +95,14 @@ instance Arbitrary PV1.Data where
     where
       gendata n
         | n > 0 =
-          oneof
-            [ PV1.I <$> arbitrary,
-              PV1.B <$> arbitrary,
-              PV1.Map <$> listOf ((,) <$> gendata (n `div` 2) <*> gendata (n `div` 2)),
-              PV1.Constr <$> fmap fromIntegral (arbitrary :: Gen Natural)
-                <*> listOf (gendata (n `div` 2)),
-              PV1.List <$> listOf (gendata (n `div` 2))
-            ]
+            oneof
+              [ PV1.I <$> arbitrary,
+                PV1.B <$> arbitrary,
+                PV1.Map <$> listOf ((,) <$> gendata (n `div` 2) <*> gendata (n `div` 2)),
+                PV1.Constr <$> fmap fromIntegral (arbitrary :: Gen Natural)
+                  <*> listOf (gendata (n `div` 2)),
+                PV1.List <$> listOf (gendata (n `div` 2))
+              ]
       gendata _ = oneof [PV1.I <$> arbitrary, PV1.B <$> arbitrary]
 
 instance
@@ -123,11 +112,7 @@ instance
   ) =>
   Arbitrary (AlonzoAuxiliaryData era)
   where
-  arbitrary =
-    frequency
-      [ (9, AlonzoAuxiliaryData <$> arbitrary <*> arbitrary),
-        (1, pure $ AlonzoAuxiliaryData mempty mempty)
-      ]
+  arbitrary = AlonzoAuxiliaryData <$> arbitrary <*> arbitrary
 
 instance Arbitrary Tag where
   arbitrary = elements [Spend, Mint, Cert, Rewrd]
@@ -413,6 +398,7 @@ instance
       [ pure NoDatum,
         DatumHash <$> arbitrary,
         Datum . dataToBinaryData <$> arbitrary
+      ]
 
 instance (Era era, Val (Value era), DecodeNonNegative (Value era)) => Twiddle (AlonzoTxOut era) where
   twiddle = twiddle . toTerm
@@ -437,9 +423,7 @@ emptyOrNothing ::
     Twiddle (t a)
   ) =>
   AlonzoTxBody (AlonzoEra crypto) ->
-  ( AlonzoTxBody (AlonzoEra crypto) ->
-    t a
-  ) ->
+  (AlonzoTxBody (AlonzoEra crypto) -> t a) ->
   Gen (Maybe Term)
 emptyOrNothing txBody f =
   if null $ f txBody
@@ -478,7 +462,6 @@ instance Twiddle Coin where
 instance C.Crypto crypto => Twiddle (AlonzoTxBody (AlonzoEra crypto)) where
   twiddle txBody = do
     inputs' <- twiddle $ inputs txBody
-    traceM $ "inputs': " <> (show  . toLazyByteString $ encodeTerm inputs')
     outputs' <- twiddle $ outputs txBody
     fee' <- twiddle $ txfee txBody
     -- Empty collateral can be represented by empty set or the
@@ -490,12 +473,12 @@ instance C.Crypto crypto => Twiddle (AlonzoTxBody (AlonzoEra crypto)) where
     auxDataHash' <- twiddleStrictMaybe $ adHash txBody
     validityStart' <- twiddleStrictMaybe . invalidBefore $ txvldt txBody
     mint' <- twiddle $ mint txBody
-    scriptDataHash' <- twiddleStrictMaybe $ scriptIntegrityHash txBody 
+    scriptDataHash' <- twiddleStrictMaybe $ scriptIntegrityHash txBody
     collateral' <- emptyOrNothing txBody collateral
     requiredSigners' <- emptyOrNothing txBody reqSignerHashes
     networkId' <- twiddleStrictMaybe $ txnetworkid txBody
     mp <- elements [TMap, TMapI]
-    let fields = 
+    let fields =
           [ (TInt 0, inputs'),
             (TInt 1, outputs'),
             (TInt 2, fee')