+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Serialisation/EraIndepGenerators.hs
@@ -147,7 +147,7 @@ import Test.QuickCheck
     shrink,
     vectorOf,
   )
-import Test.QuickCheck.Gen (chooseAny, frequency)
+import Test.QuickCheck.Gen (chooseAny)
 
 -- =======================================================
 
@@ -649,7 +649,7 @@ instance Arbitrary Url where
   arbitrary = return . fromJust $ textToUrl "text"
 
 instance Arbitrary a => Arbitrary (StrictSeq a) where
-  arbitrary = StrictSeq.forceToStrict <$> frequency [(9, arbitrary), (1, pure mempty)]
+  arbitrary = StrictSeq.forceToStrict <$> arbitrary
   shrink = map StrictSeq.forceToStrict . shrink . StrictSeq.fromStrict
 
 instance Arbitrary StakePoolRelay where